requirements.txt
adrf==0.1.6
aenum==3.1.16
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==4.4.0
APScheduler==3.10.1
asgiref==3.7.2
async-property==0.2.2
attrs==23.2.0
awscli==1.35.20
awsebcli==3.21.0
blessed==1.22.0
boto3==1.35.54
botocore==1.35.54
cashfree-pg==4.3.10
cement==2.10.14
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.5.8
Django==5.0.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
django-quill-editor==0.1.40
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docutils==0.16
dotsi==0.0.3
frozenlist==1.4.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
jmespath==1.0.1
Markdown==3.5.2
marshmallow==3.26.1
marshmallow-enum==1.5.1
multidict==6.0.5
mypy_extensions==1.1.0
packaging==24.0
pathspec==0.10.1
phonenumbers==8.13.46
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==1.10.24
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.3
requests==2.31.0
rsa==4.7.2
s3transfer==0.10.1
semantic-version==2.10.0
sentry-sdk==1.32.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
termcolor==2.5.0
twilio==9.0.0
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2024.1
tzlocal==5.3.1
urllib3==1.26.12
uvicorn==0.30.1
wcwidth==0.2.14
yarl==1.9.4
